# Documentation/openapi.yaml
openapi: 3.0.0
info:
  title: Supchat
  description: une API de chat pour une entreprise
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Serveur local Symfony
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Channels:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 3
          maxLength: 255
        workspace:
          $ref: '#/components/schemas/Workspaces'
        status:
          type: boolean
          description: true = public, false = private
        minRole:
          type: integer
          default: 1
          description: 1 = membre, 2 = modérateur, 3 = admin
      required: [name, workspace, status]
...
    Users:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
          minLength: 3
          maxLength: 255
        lastName:
          type: string
          minLength: 3
          maxLength: 255
        userName:
          type: string
          minLength: 3
          maxLength: 255
        emailAddress:
          type: string
          format: email
          minLength: 3
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 255
          nullable: true
        role:
          type: string
          enum: ["ROLE_USER", "ROLE_ADMIN"]
        theme:
          type: boolean
        status:
          type: string
          minLength: 3
          maxLength: 255
      required: [firstName, lastName, userName, emailAddress, role, theme, status]
...
    Workspaces:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 3
          maxLength: 255
        status:
          type: boolean
          description: true = public, false = private
        creator:
          $ref: '#/components/schemas/Users'
      required: [name, status, creator]
...
paths:
...
  /api/channels/{id}:
    get:
      tags: [Channels]
      summary: Récupère un channel
      operationId: getChannel
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détails du channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channels'
        '403':
          description: Accès interdit
        '404':
          description: Channel introuvable

    put:
      tags: [Channels]
      summary: Met à jour un channel
      operationId: updateChannel
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  type: boolean
                min_role:
                  type: integer
      responses:
        '200':
          description: Canal mis à jour
        '400':
          description: Requête invalide

    delete:
      tags: [Channels]
      summary: Supprime un channel
      operationId: deleteChannel
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Canal supprimé
...
  /api/messages:
    post:
      tags: [Messages]
      summary: Crée un nouveau message
      operationId: createMessage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [channel_id, user_id, content]
              properties:
                channel_id:
                  type: integer
                user_id:
                  type: integer
                content:
                  type: string
      responses:
        '201':
          description: Message créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  content: { type: string }
                  timestamp: { type: string, format: date-time }
                  author:
                    type: object
                    properties:
                      id: { type: integer }
                      username: { type: string }
                  channel_id: { type: integer }
        '400':
          description: Données incomplètes
        '404':
          description: Channel ou utilisateur non trouvé
  /api/messages/{id}:
    get:
      tags: [Messages]
      summary: Récupère un message
      operationId: getMessage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Message trouvé
        '404':
          description: Message introuvable

    put:
      tags: [Messages]
      summary: Met à jour le contenu d’un message
      operationId: updateMessage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [content]
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Message modifié
        '400':
          description: Contenu invalide

    delete:
      tags: [Messages]
      summary: Supprime un message
      operationId: deleteMessage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Message supprimé
  /api/channels/{id}/messages:
    get:
      tags: [Messages]
      summary: Récupère les messages d’un canal
      operationId: getChannelMessages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Liste des messages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    content: { type: string }
                    timestamp: { type: string, format: date-time }
                    author:
                      type: object
                      properties:
                        id: { type: integer }
                        username: { type: string }
                    channel_id: { type: integer }
                    hashtags:
                      type: array
                      items:
                        type: object
                        properties:
                          tag: { type: string }
                          channel:
                            type: object
                            properties:
                              id: { type: integer }
                              name: { type: string }
                    files:
                      type: array
                      items:
                        type: object
                        properties:
                          id: { type: integer }
                          name: { type: string }
                          download_url: { type: string }
        '404':
          description: Canal non trouvé
...